tests: switch to uv pip-compile for lockfiles
This switches the nox pip-compile session and lockfiles to use uv pip
compile (written in Rust and much faster) instead of pip-compile from
pip-tools for our dependency update jobs.

As a side effect, this addresses the issue brought up in
https://github.com/ansible/ansible-documentation/pull/1950#issuecomment-2443965477,
as we're no longer using pip-compile that can break anytime pip changes
the private/internal APIs that pip-tools relies on.

I re-generated the lockfiles with `--no-upgrade` to avoid extraneous
changes in this PR.
The only additions are the formatting of comments in the lockfile and
colorama.
colorama was added as a result of `uv pip-compile`'s `--universal` flag
which creates "universal" lockfiles that include dependencies of other
platforms, architectures, and Python versions than those of the system
used to generate the lockfile.
`--universal` should create more consistent results for contributors who
use other systems
(e.g., Mac OS or a newer Python version than the one used in CI).

Fixes: https://github.com/ansible/ansible-documentation/issues/1755